Create README.md
The main idea of this project is to provide easiness for a client who is interested in building a house, office, shops etc. directly from his/her phone. The android-based application will allow the user to get an estimate of the house of his dreams. Direct contacts to the top contractors in town and material providers (if interested in self-construction) all through a simple touch of finger. Estimation of cost needed to build the house, Contact information of top contractors along with their profile, ratings, and updated material prices that are prevailing in the market is our main priority.
We are going to develop an app that is beneficial for both users as well as contractors. Users can find the best Estimation of Expenditure to construct a house whereas contractors can display their services to user. Users can select the best contractor.
Users can give reviews to the contractors whether they have bad or good experience with them. We will also provide updated material prices so that user can compare that prices with the prevailing market prices and after comparison, he will be in a position to decide in a better way and can save his money. 
After getting idea, he will in a position to argue with the contractors and he can make him liable for over costing. Ultimately, the contractors will provide best of their services to their customers to avoid any inconvenience. It will benefit the construction industry at large.